process.py: add support to spawn a multithreaded process
Add support for creating a multithreaded process. The multithreaded
process gets created by spawning a new process, that in turn creates
N threads. The thread function is a while loop, that sleeps for N
seconds. Currently, there is no support for waiting for the threads
to join, and to stop the threads, the spawned process needs to be
terminated/killed.

Suggested-by: Tom Hromatka <tom.hromatka@oracle.com>
Signed-off-by: Kamalesh Babulal <kamalesh.babulal@oracle.com>