github: Run the libcgroup Github Actions
Add jobs to run the libcgroup automated tests against the main
branch of libcgroup.  Since the Github Action is being
triggered from the tests repo, the setup steps differ slightly
from libcgroup's steps.  In this case we need to move the unit
and functional tests from the root to tests/, and we need to
move the library code from libcg/ to the root.

Finally, code coverage is not enabled in these runs.

Signed-off-by: Tom Hromatka <tom.hromatka@oracle.com>