test-019: fix issues on unsuccessful no-container runs
Running the test with --no-container, might cause an issue with useradd
or groupadd command if the home directory of the test user exists/test
user group exists.

The home directory of the user doesn't get removed by default on all
distros, so explicitly pass the '-r' flag to userdel to remove the
directory.  Whereas user group might exist due to unsuccessful test
runs, address it by passing '-f' flags to groupadd command, that
returns success even if the group exists, that will be removed
by the groupdel command at the exit.

This is not an issue with containers test cases, because they are
created and destroyed for every test case execution.

Signed-off-by: Kamalesh Babulal <kamalesh.babulal@oracle.com>