gunit: add installcheck
Add support for installing testcases, so that the users can run it later
after installation too.  The installcheck hook is part of GNU automake
and under the hook, gtest binary gets build and copied over to
$(libdir)/libcgroup/tests/, along with its dependent library
googletest/libgtest.so.

Currently, the gtest is built/run with make check. While building it
uses -rpath to link with <libcgroup source>/googletest/../libgtest.so.
Just copying over the binary/library might fail to run, if the libcgroup
sources are removed, to address this issue an extra -rpath of install
the destination is pre-pended during linking.

Fixes: https://github.com/libcgroup/libcgroup/issues/60

Suggested-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Kamalesh Babulal <kamalesh.babulal@oracle.com>